added Login feature and route protection for users only

+++ frontend/src/components/WorkoutDetails.js
@@ -1,11 +1,23 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
+
+// date fns
+import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
 
@@ -13,17 +25,24 @@ const WorkoutDetails = ({ workout }) => {
       dispatch({ type: "DELETE_WORKOUT", payload: json });
     }
   };
+
   return (
     <div className="workout-details">
-      <h3>{workout.title}</h3>
+      <h4>{workout.title}</h4>
       <p>
-        <strong>Load (kg): {workout.load}</strong>
+        <strong>Load (kg): </strong>
+        {workout.load}
       </p>
       <p>
-        <strong>Reps : {workout.reps}</strong>
+        <strong>Reps: </strong>
+        {workout.reps}
       </p>
-      <p>{workout.createdAt}</p>
-      <span onClick={handleClick}>delete</span>
+      <p>
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+      </p>
+      <span className="material-symbols-outlined" onClick={handleClick}>
+        delete
+      </span>
     </div>
   );
 };
